+++ app/src/main/java/com/example/weacon/Request/Call.kt
@@ -1,6 +1,6 @@
 package com.example.weacon.Request
 
-suspend fun loadWeather(lat: Int, lon: Int): WeatherObject? {
+suspend fun loadWeather(lat: Double, lon: Double): WeatherObject? {
     try {
         return GitApi.retrofitService.weatherRequest(lat.toString(), lon.toString())
     } catch(e: retrofit2.HttpException) {
